+++ src/kit/ui/UiTexts.cpp
@@ -1,67 +1,46 @@
+#include <utility>
+
 #include <stdarg.h>
 #include "UiTexts.hpp"
 
 UiTexts::UiTexts() {
-    this->init("Roboto");
+    family = "Roboto";
 }
 
 UiTexts::UiTexts(UiTheme *theme) {
     this->theme = theme;
-    this->init("Roboto");
+    this->family = "Roboto";
 }
 
 UiTexts::UiTexts(std::string family) {
-    this->init(std::move(family));
+    this->family = family;
 }
 
 UiTexts::UiTexts(std::string family, UiTheme *theme) {
     this->theme = theme;
-    this->init(std::move(family));
+    this->family = family;
 }
 
 UiTexts::~UiTexts() {
-    for (const auto& kv : this->fonts) {
-        vita2d_free_font(kv.second);
-    }
-}
-
-void UiTexts::init(std::string family) {
-    SceIoDirent gDir;
-    std::string fontsPath = DEFAULT_FONTS_PATH;
-    fontsPath = fontsPath + family;
-
-    int fd = sceIoDopen(fontsPath.c_str());//open fonts folder
-
-    std::string fontPath, fileName, fontName;
-
-    fontsPath.append("/");//add slash for folder
-
-    while (sceIoDread(fd, &gDir) > 0) {
-        fileName = gDir.d_name;
-        fontPath = fontsPath + fileName;//get the path of the file
-
-        fontName = fileName.erase(0, family.length() + 1);//remove the family name and -
-        fontName = fontName.erase(fontName.length() - 4, fontName.length());//remove .ttf
-        fonts[fontName] = vita2d_load_font_file(fontPath.c_str());
-
-    }
-
-    sceIoClose(fd);
+    this->cleanFonts();
 }
 
 void UiTexts::drawFinal(int x, int y, TextStyle textStyle, unsigned int color, bool italic, std::string text) {
     this->calcTextStyleData(textStyle, italic);
+    this->loadFont(textStyleData.type, textStyleData.size);
 
-    vita2d_font_draw_text(fonts[textStyleData.type], x, (int) floor(y + textStyleData.size - textStyleData.offset), color, textStyleData.size, textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str());
+    vita2d_font_draw_text(fonts[keyFont], x, (int) floor(y + textStyleData.size - textStyleData.offset), color, textStyleData.size, textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str());
 }
 
 void UiTexts::drawFinal(int x, int y, TextStyleData _textStyleData, unsigned int color, std::string text) {
-    vita2d_font_draw_text(fonts[_textStyleData.type], x, (int) floor(y + _textStyleData.size - _textStyleData.offset), color, _textStyleData.size, _textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str());
+    this->loadFont(_textStyleData.type, _textStyleData.size);
+
+    vita2d_font_draw_text(fonts[keyFont], x, (int) floor(y + _textStyleData.size - _textStyleData.offset), color, _textStyleData.size, _textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str());
 }
 
 //Draw with Material Style
 void UiTexts::draw(int x, int y, TextStyle textStyle, std::string text) {
-    this->drawFinal(x, y, textStyle, DEFAULT_FONT_COLOR, false, std::move(text));
+    this->drawFinal(x, y, textStyle, TEXTS_DEFAULT_FONT_COLOR, false, std::move(text));
 }
 
 void UiTexts::draw(int x, int y, TextStyle textStyle, unsigned int color, std::string text) {
@@ -106,7 +85,7 @@ void UiTexts::drawF(int x, int y, TextStyle textStyle, TypeTheme textThemeColor,
 
 //Draw with your style
 void UiTexts::draw(int x, int y, TextStyleData _textStyleData, std::string text) {
-    this->drawFinal(x, y, _textStyleData, (unsigned int) DEFAULT_FONT_COLOR, std::move(text));
+    this->drawFinal(x, y, _textStyleData, TEXTS_DEFAULT_FONT_COLOR, std::move(text));
 }
 
 void UiTexts::draw(int x, int y, TextStyleData _textStyleData, unsigned int color, std::string text) {
@@ -122,6 +101,17 @@ void UiTexts::drawF(int x, int y, TextStyleData _textStyleData, unsigned int col
     va_end(argPtr);
 
     this->drawFinal(x, y, _textStyleData, color, buf);
+}
+
+void UiTexts::drawF(int x, int y, TextStyleData _textStyleData, TypeTheme typeTheme, const char *text, ...) {
+    char buf[1024];
+    va_list argPtr;
+
+    va_start(argPtr, text);
+    vsnprintf(buf, sizeof(buf), text, argPtr);
+    va_end(argPtr);
+
+    this->drawFinal(x, y, _textStyleData, typeTheme == THEME_PRIMARY ? this->theme->getPrimaryRGBA().text : this->theme->getSecondaryRGBA().text, buf);
 }
 
 
@@ -140,16 +130,17 @@ std::string UiTexts::toUppercase(std::string text) {
 }
 
 void UiTexts::calcTextData(std::string text, TextStyle textStyle, bool italic) {
-
     this->calcTextStyleData(textStyle, italic);
+    this->loadFont(textStyleData.type, textStyleData.size);
 
-    vita2d_font_text_dimensions(fonts[textStyleData.type], textStyleData.size, textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str(), &textData.width, &textData.height);
+    vita2d_font_text_dimensions(fonts[keyFont], textStyleData.size, textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str(), &textData.width, &textData.height);
 }
 
 
 TextData UiTexts::getTextData(std::string text, TextStyleData _textStyleData) {
+    this->loadFont(textStyleData.type, textStyleData.size);
 
-    vita2d_font_text_dimensions(fonts[_textStyleData.type], _textStyleData.size, _textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str(), &textData.width, &textData.height);
+    vita2d_font_text_dimensions(fonts[keyFont], _textStyleData.size, _textStyleData.uppercase ? this->toUppercase(text).c_str() : text.c_str(), &textData.width, &textData.height);
 
     return textData;
 }
@@ -212,7 +203,43 @@ void UiTexts::calcTextStyleData(TextStyle textStyle, bool italic) {
             break;
 
     }
-    textStyleData.size += DEFAULT_SIZE_OFFSET;
+    textStyleData.size += TEXTS_DEFAULT_SIZE_OFFSET;
 
     textStyleData.offset = textStyleData.size * 0.24;//correct font position
 }
+
+//dynamic import
+std::pair<std::string, int> UiTexts::loadFont(std::string type, int size) {
+    keyFont = std::make_pair(type, size);
+
+    if (fonts.find(keyFont) == fonts.end()) {
+
+        fontPath = TEXTS_DEFAULT_FONTS_PATH;
+        fontPath.append(family);
+        fontPath.append("/");
+        fontPath.append(family);
+        fontPath.append("-");
+        fontPath.append(type);
+        fontPath.append(".ttf");
+
+        fonts[keyFont] = vita2d_load_font_file(fontPath.c_str());
+    }
+
+    return keyFont;
+}
+
+void UiTexts::cleanFonts() {
+    for (const auto& kv : this->fonts) {
+        vita2d_free_font(kv.second);
+    }
+    fonts.clear();
+}
+
+void UiTexts::cleanFont(std::string type, int size) {
+    keyFont = std::make_pair(type, size);
+    if (fonts.find(keyFont) == fonts.end()) {
+        vita2d_free_font(fonts[keyFont]);
+    }
+    fonts.erase(keyFont);
+}
+